Adding the query inside the graph
